refactor(preact): move install prompt to preact
